Atom states moved to data file
